requirements.txt
python-barcode
openpyxl
reportlab